last updated 02.13.2023

+++ Back_end/Java/Spring/BookClub/src/main/java/com/alex/bookclub/services/BookService.java
@@ -27,7 +27,13 @@ public class BookService {
 		}
 		return null;
 	}
-	public void deleteBookById(Long id) {
-		repo.deleteById(id);
+	public Book deleteBookById(Long id) {
+		Optional<Book> book = repo.findById(id);
+		if(book.isPresent()) {
+			repo.deleteById(id);
+		} else {
+			System.out.println("error");
+		}
+		return null;
 	}
 }
